--- a/app/views/support/cases/message_threads/stream_events/_marked_as_read.turbo_stream.erb
+++ b/app/views/support/cases/message_threads/stream_events/_marked_as_read.turbo_stream.erb
@@ -2,6 +2,11 @@
   <%= render "support/cases/message_threads/message", message: %>
 <% end %>
 
-<%= turbo_stream.replace("action-flag") do %>
-  <%= render "support/cases/action_flag", badge: true if kase.action_required? && !kase.closed? %>
+<% if kase.action_required? && !kase.closed? %>
+  <%= turbo_stream.append("case-badges") do %>
+    <%= render "support/cases/action_flag", badge: true %>
+  <% end %>
+<% else %>
+  <%= turbo_stream.remove("action-flag") do %>
+  <% end %>
 <% end %>