--- a/app/views/support/cases/show/_tabs.html.erb
+++ b/app/views/support/cases/show/_tabs.html.erb
@@ -11,7 +11,7 @@
 
 <%= turbo_stream_from "case_status_updates" %>
 
-<div class="flex-gap-5px govuk-!-margin-bottom-7">
+<div id="case-badges" class="flex-gap-5px govuk-!-margin-bottom-7">
   <%= render "support/cases/status_badge", state: @current_case.state %>
   <%= render "support/cases/with_school_badge" if @current_case.with_school %>
   <%= render "support/cases/accessibility_flag" if @current_case.has_special_requirements? %>