--- a/db/migrate/20240703155237_add_support_email_attachments_indexes.rb
+++ b/db/migrate/20240703155237_add_support_email_attachments_indexes.rb
@@ -1,0 +1,7 @@
+class AddSupportEmailAttachmentsIndexes < ActiveRecord::Migration[7.1]
+  disable_ddl_transaction!
+
+  def change
+    add_index :support_email_attachments, :email_id, algorithm: :concurrently
+  end
+end