## Changes in this PR
### Performance issues

After trying out different combinations of indexes for the `support_emails` table and inspecting the query plan outputted by `EXPLAIN ANALYSE`, I've found that the Postgres planner seems to prefer the 3 indexes I've added to the schema.

Tested on local data only (significantly smaller number of rows than production), I see these reductions in **execution time**:
- for the`support_message_threads` view, down from ~180 ms to 15-20 ms
- for a query that fetches all the threads for a case with 35 threads and 64 emails (where `ticket_id` and `ticket_type` are specified in the `WHERE` clause), down from ~20 ms to ~1.5 ms
- for a query that fetches all the emails for a case with 64 emails (where `ticket_id` and `ticket_type` are specified in the `WHERE` clause), down from ~0.4 ms to ~0.1 ms

`support_email_attachments` also gets an index on `email_id`, which should speed up the loading of attachments when viewing an email. Postgres didn't seem to take advantage of multi-column indexes here, though we do use `hidden` and `is_inline` in our queries.

### Bugs

Fixed a bug where `turbo-stream` was not reloading messages to show their new read/unread status after clicking "Mark as read" or "Mark as unread" in a thread. This was due to the controller initialising the email as a `Support::Email` whereas the messages in the thread view are initialised as `Email`, thus resulting in a different ID when calling `dom_id` and preventing `turbo-stream` from locating the `div` for the relevant message.